move export parsing to seperate module, started on export doc-pages
