+++ rnote-cli/src/main.rs
@@ -3,6 +3,7 @@
 //! The cli interface is not (yet) stable and could change at any time.
 
 pub(crate) mod cli;
+pub(crate) mod export;
 
 fn main() -> anyhow::Result<()> {
     smol::block_on(async { cli::run().await })
