+++ rnote-cli/src/meson.build
@@ -2,4 +2,5 @@
 rnote_cli_sources = files(
     'main.rs',
     'cli.rs',
+    'export.rs',
 )
